Move Adapter, MappingAdapter to .util.common
